Publish / Subscriber
Esqueleto de publicador suscriptor

+++ controller_node/mavros/torpi/src/publisher_gps_fotos.h
@@ -31,7 +31,7 @@ public:
         //pthread create (thread-parameters, blabla, this->run(), blabla)
     }
 
-    void publish(int latitud, int longitud, int altitud);
+    void publish(double latitud, double longitud, double altitud);
 
     void publish();
